requirements.txt
google-api-python-client ~= 2.124.0
google-auth-httplib2 ~= 0.2.0
google-auth-oauthlib ~= 1.2.0
beautifulsoup4 ~= 4.12.3
PyYAML ~= 6.0.1
psycopg2 ~= 2.9.9
